test(Line): 리팩토링
- LineAcceptanceTest, LineSectionAcceptanceTest 병합
- 메소드명 변경 및 리팩토링
