+++ src/test/java/subway/fixtures/LineFixtures.java
@@ -32,4 +32,12 @@ public class LineFixtures {
         return params;
     }
 
+    public static Map<String, String> 구간_등록_파라미터_생성(String upStationId, String downStationId) {
+        Map<String, String> params = new HashMap<>();
+        params.put("upStationId", upStationId);
+        params.put("downStationId", downStationId);
+        params.put("distance", "10");
+        return params;
+    }
+
 }
